Rename Java Plugin to JavaPlugin
