+++ JavaPlugin/editor/api/Plugin.class
