+++ JavaPlugin/editor/app/Editor$1.class
