+++ JavaPlugin/editor/app/Editor$2.class
