+++ JavaPlugin/editor/app/Editor$3.class
