+++ JavaPlugin/editor/app/Editor.class
