+++ JavaPlugin/plugins/touppercase/ToUppercasePlugin.class
