+++ JavaPlugin/plugins/wordcount.jar
