+++ JavaPlugin/plugins/wordcount/WordCountPlugin.class
